final norminette for my part
